fix: api로직 따로 분리
